Improving parse and unsimplifying namespace
The book demonstrates a nice shortcut bijection for testing divisibility
by 3 in base 10: a number is evenly divisible by 3 iff the sum all of
all its base-10 digits is evenly divisible by 3. I have a notion that
this generalizes to divisibility by n in base n^2 + 1, which I'd like to
test on base 5 (2^2 + 1), base 17 (4^2 + 1), base 26 (5^2 + 1), etc.

I need to work on the names for the arbitrary base functions, but this
commit starts to build up the functions necessary to convert between
arbitrary base and Clojure numerics so that I can test this theory.